isolate wallet code into separate bundles
