--- wormhole-connect/src/utils/sei.ts
@@ -1,4 +1,0 @@
-import { Network } from '@certusone/wormhole-sdk';
-
-export const getSeiChainId = (env: Network) =>
-  env === 'MAINNET' ? 'pacific-1' : 'atlantic-2';
