+++ wormhole-connect/src/views/Bridge/Inputs/AmountInput.tsx
@@ -62,6 +62,7 @@ function AmountInput(props: Props) {
       label={props.label ?? 'Amount'}
       error={!!(showErrors && validations.amount)}
       editable
+      disabled
       onClick={focus}
       cursor="text"
     >
