+++ wormhole-connect/src/views/Redeem/SendTo.tsx
@@ -16,8 +16,8 @@ import RouteOperator from 'routes/operator';
 import {
   TransferWallet,
   registerWalletSigner,
-  signSolanaTransaction,
   switchChain,
+  signAndSendTransaction,
 } from 'utils/wallet';
 
 import AlertBanner from 'components/AlertBanner';
@@ -51,7 +51,7 @@ function AssociatedTokenAlert() {
     );
     // if `tx` is null it means the account already exists
     if (!tx) return;
-    await signSolanaTransaction(tx, TransferWallet.RECEIVING);
+    await signAndSendTransaction('solana', tx, TransferWallet.RECEIVING);
     dispatch(setTxDetails({ ...txData, recipient: wallet.address }));
   }, [txData, wallet.address, dispatch]);
